Move class to header file
